Add file via upload
Fib() requires an input n between 0~51,  which is the length of Fibonacci sequence it creates.  
You can put it in the for loop just like an iteration, it will return the value less than 100000 among the infinite Fibonacci sequence.  
You can also treat it like a list using slice, it will return the result you want among the Fibonacci sequence it just created.